feat: implement the function to receive the input about bonus number and validate that the input is proper
