+++ README.md
@@ -44,13 +44,14 @@ Lotto 클래스의 생성자에 Collections.sort() 메소드를 이용함.
   3. 입력받은 숫자들이 1 ~ 45 사이의 수인가?
   4. 입력받은 숫자들이 서로 중복되지 않는가?
 
-- \[ ]보너스 번호를 입력받는 함수
-- \[ ]입력받은 보너스 번호가 올바른 값인지 검증하는 함수
+- \[X]보너스 번호를 입력받는 함수
+- \[X]입력받은 보너스 번호가 올바른 값인지 검증하는 함수
   1. 입력받은 값이 숫자가 맞는가?
   2. 입력받은 숫자가 1개가 맞는가?  
 => Integer.parseInt 메소드로 변환 이후, 1 ~ 45 사이의 수인지에 대한 검증을 통해 묵시적으로 검증된다.
   3. 입력받은 숫자가 1 ~ 45 사이의 수인가?
-  4. 입력받은 수가 먼저 입력받은 당첨 번호와 중복되지 않는가?
+  4. 입력받은 수가 먼저 입력받은 당첨 번호와 중복되지 않는가?  
+=> 11월 13일 새벽 1시 48분 기준, 좋은 수가 떠오르지 않아서 main() 함수에서 직접 검증하도록 구현하였다.
 
 - \[ ]발행한 로또들과, 입력받은 당첨 번호 및 보너스 번호를 비교하여 당첨 여부(나아가 당첨 등수)를 검사하는 함수
 - \[ ]당첨 여부를 바탕으로 당첨 내역을 출력하는 함수
